wip: update image
