+++ dev/test-studio/config/.checksums
@@ -1,0 +1,9 @@
+{
+  "#": "Used by Sanity to keep track of configuration file checksums, do not delete or modify!",
+  "@sanity/default-layout": "bb034f391ba508a6ca8cd971967cbedeb131c4d19b17b28a0895f32db5d568ea",
+  "@sanity/default-login": "6fb6d3800aa71346e1b84d95bbcaa287879456f2922372bb0294e30b968cd37f",
+  "@sanity/form-builder": "b38478227ba5e22c91981da4b53436df22e48ff25238a55a973ed620be5068aa",
+  "@sanity/data-aspects": "d199e2c199b3e26cd28b68dc84d7fc01c9186bf5089580f2e2446994d36b3cb6",
+  "@sanity/google-maps-input": "43e62bbd0a78d3dfd363915959160d2dfeafbac941d5a78a0a4c4fd64d9e7638",
+  "mux-input": "595f2707bb80b1584160b3e4aa64454d9f89d1c5ce0f40b4de718d02fd1b3744"
+}
