+++ dev/test-studio/config/@sanity/default-login.json
@@ -1,0 +1,7 @@
+{
+  "providers": {
+    "mode": "append",
+    "redirectOnSingle": false,
+    "entries": []
+  }
+}
